requirements.txt
autopep8==1.5.7
click==7.1.2
gmpy2==2.0.8
pycodestyle==2.8.0
pycryptodome==3.11.0
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
python-dotenv==0.19.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
toml==0.10.2
